Reduce the histogram bins to 32-bits
The 64-bit bins were overkill as if you need them to be that large
the counts variable would likely overflow. By reducing to 32-bit
bins the histograms can be up to 32GB before the counts variable
could possibly overflow.

+++ src/active_info.c
@@ -101,7 +101,7 @@ double inform_active_info(int const *series, size_t n, size_t m, int b, size_t k
     size_t const futures_size = b;
     size_t const total_size = states_size + histories_size + futures_size;
 
-    uint64_t *data = calloc(total_size, sizeof(uint64_t));
+    uint32_t *data = calloc(total_size, sizeof(uint32_t));
     if (data == NULL)
     {
         return inform_nan(7);
@@ -172,7 +172,7 @@ int inform_local_active_info(int const *series, size_t n, size_t m,
     size_t const futures_size = b;
     size_t const total_size = states_size + histories_size + futures_size;
 
-    uint64_t *data = calloc(total_size, sizeof(uint64_t));
+    uint32_t *data = calloc(total_size, sizeof(uint32_t));
     if (data == NULL)
     {
         return 6;
